# Refs #27624 -- Optimized sql.Query creation by moving immutable/singleton attributes to class attributes.