@charettes 2021-11-25
Small note that this was attempted a few years ago and dismissed because it was reported that it would slow down attribute accesses too much https://github.com/django/django/pull/9597#issuecomment-358933789.

If we have numbers backing up that this is also an improvement for methods that do a lot of attribute accesses then I think it's worth merging and maybe doing the same with `QuerySet`? I'd try expanding your set of tests to perform joins, aggregation, rely on `values`, `annotate`, `select_related`, force grouping with `values.annotate` as these operations are more _read_ intensive than the benchmarks you used. Maybe [`djangobench`](https://github.com/django/djangobench) could be used as a good exploration tool as well.

Also, benchmarks that do I/O and see drastic improvements (e.g. reaches out to the database such as `tuple(User.objects.prefetch_related('groups', 'user_permissions'))`) make me skeptical as I would expect the time saved in attributes access to be dwarfed by the time spent doing I/O.