@kezabelle 2021-11-25
Hey Simon, thanks so much for taking a look.

> Small note that this was attempted a few years ago and dismissed because it was reported that it would slow down attribute accesses too much [#9597 (comment)](https://github.com/django/django/pull/9597#issuecomment-358933789).

That's useful, I'll have to take a look at it, thanks so much for digging it out for me.

> If we have numbers backing up that this is also an improvement for methods that do a lot of attribute accesses then I think it's worth merging and maybe doing the same with `QuerySet`? 

Yeah, _if_ this works out (and it's a bigger _if_ based on the above prior attempt), I've already got patches for `QuerySet` and `ManyRelatedManager` and `GenericRelatedObjectManager` in place (and _maybe_ `Manager` proper).

> I'd try expanding your set of tests to perform joins, aggregation, rely on `values`, `annotate`, `select_related`, force grouping with `values.annotate` as these operations are more _read_ intensive than the benchmarks you used. Maybe [`djangobench`](https://github.com/django/djangobench) could be used as a good exploration tool as well.

A fair idea; perhapsI can find the availability to do so.

> Also, benchmarks that do I/O and see drastic improvements (e.g. reaches out to the database such as `tuple(User.objects.prefetch_related('groups', 'user_permissions'))`) make me skeptical as I would expect the time saved in attributes access to be dwarfed by the time spent doing I/O.

Most of the time spent is _not_ I/O, that's the _problem_ with `prefetch_related` and why it has been my target for a few attempted patches:
```
In [2]: x = set(range(1, 1001))
In [3]: %timeit tuple(User.objects.filter(pk__in=x))
13.1 ms ± 318 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
In [4]: connection.queries[-1]['time']
Out[4]: '0.001'
```
20~% of that chunk is inflating `User` model objects, and 10% running through `apply_converters` which is unavoidable (the latter being SQLite specific, so call it 20% for the other backends).

Doing the same for `groups`:
```
In [5]: Group.objects.count()
0 
In [6]: %timeit tuple(Group.objects.filter(pk__in=x))
2.14 ms ± 30.9 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
In [7]: connection.queries[-1]['time']
Out[7]: '0.001'
```

The query for `user_permissions` is more complex to bother to try and set up off-the-cuff, but being that there are **0** user permissions (`User.user_permissions.through.objects.count()`) and the reported value from `connection.queries` is again `0.001` I think it's _reasonable_ to assume that it's either 3 or perhaps even 15ms.

In total that's only *30ms* of time to execute the queries + return *1000* `User` instances. All the other time is basically book-keeping by Django/Python to actually facilitate that. It's _that_ overhead that I'm trying to reduce, precisely because it's not I/O. Much of it is generating related managers for each instance in each level of the prefetch tree, hence things like #14724 and #14675 to try and reduce the costs of those.

```
In [2]: %timeit tuple(User.objects.prefetch_related(None))
11.1 ms ± 329 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [3]: %timeit tuple(User.objects.prefetch_related('groups'))
48.9 ms ± 1.83 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [4]: %timeit tuple(User.objects.prefetch_related('user_permissions'))
49.3 ms ± 662 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [5]: %timeit tuple(User.objects.prefetch_related('groups', 'user_permissions'))
87.6 ms ± 991 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

All of the above give me the same results; all of them should in an ideal world be as close to prefetching `None` as possible, though naturally _some_ book-keeping and execution _has_ to occur ;)