@kezabelle 2021-11-26
Adding as a second comment for separation and clarity.

> Small note that this was attempted a few years ago and dismissed because it was reported that it would slow down attribute accesses too much.

It's a shame that no numbers were present in that thread, or the specific particulars being discussed around attribute access, as the difference seems negligible to me based on non-Django-specific (pure Python) behaviour. I could naturally be understanding incorrectly as a result, but:

```
class Demo:
    a = True
    def __init__(self) -> None:
        self.b = True
```

Get a baseline time of class instantiation.
Ran multiple times to demonstrate the CPU workload flux more.
```
In [2]: %timeit Demo()
211 ns ± 5.21 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
In [3]: %timeit Demo()
231 ns ± 9.2 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
In [4]: %timeit Demo()
207 ns ± 2.02 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

Assignment to a class attribute on an instance is the only bit that can be slower (AFAIK) because it's pushing a value into the instance `__dict__`; thereafter, assignments or mutations should be the same as normal instance attributes (because they are). As you can only do the conversion of class attribute to instance attribute once, it's necessary to time the creation of the instance too (AFAICT), hence the baseline above. Here's a couple of runs:

```
In [5]: %%timeit
   ...: a = Demo()
   ...: a.a = False
243 ns ± 5.8 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [6]: %%timeit
   ...: a = Demo()
   ...: a.a = False
252 ns ± 10.5 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

and here's the same runs for assigning to the _instance_ attribute with the same value:

```
In [7]: %%timeit
    ...: a = Demo()
    ...: a.b = False
239 ns ± 3.52 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [8]: %%timeit
    ...: a = Demo()
    ...: a.b = False
261 ns ± 10 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

Although we could also profile it by deleting from the instance's `__dict__` I suppose:

```
In [9]: %%timeit
   ...: a.a = False
   ...: a.__dict__.pop('a')
133 ns ± 1.11 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [10]: %%timeit
    ...: a.b = False
    ...: a.__dict__.pop('b')
130 ns ± 0.579 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
```

And then for _access_, where we can instead hold a single instance of `Demo` and just test the attributes:
```
In [11]: %timeit a.a
32.6 ns ± 0.392 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [12]: %timeit a.b
33.3 ns ± 1.44 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each
```
And then let's replace the class variable with an instance one:
```
In [13]: a.a = True
In [14]: %timeit a.a
33.3 ns ± 0.112 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
```

So at an abstract high level, I can't see much of a difference (on `Python 3.9.5`). Assuming I'm understanding the discussion correctly. I'm not going to sweat over a couple of _nanoseconds_, but if I were, I'd be doing so in a whole bunch of other places instead.


> I'd try expanding your set of tests to perform joins, aggregation, rely on `values`, `annotate`, `select_related`, force grouping with `values.annotate` as these operations are more _read_ intensive than the benchmarks you used. Maybe djangobench could be used as a good exploration tool as well.

So, I installed `djangobench` into my extant virtualenv for Django, which went fine. Then I tried to do `djangobench --control=3.2.9 --experiment=main` to get a baseline difference. And somewhere within `djangobench.main.get_django_version` it decided to **literally delete the contents of my venv folder**, such that where I previously had:
```
drwxr-xr-x  21 kez  staff  672 26 Nov 10:51 bin
drwxr-xr-x   2 kez  staff   64 26 Nov 10:50 include
drwxr-xr-x   3 kez  staff   96 26 Nov 10:50 lib
-rw-r--r--   1 kez  staff  103 26 Nov 10:50 pyvenv.cfg
drwxr-xr-x   3 kez  staff   96 26 Nov 10:51 share
drwxr-xr-x   3 kez  staff   96 26 Nov 10:51 src
```
I subsequently only had the following (which is the editable install for `djangobench`)
```
drwxr-xr-x   3 kez  staff   96 26 Nov 10:51 src
```
Those timestamps? Those are me having to re-install my virtualenv as a result. So ... yeah, I won't be running _that_. I'm considering myself lucky that it didn't wipe out more (e.g: all my experiment branches locally in git). The only output before a stacktrace (because it couldn't _find_ python, having just deleted it!) was: `Running all benchmarks`.

Anyway, as a quick first foray into other, more read intensive (I'm not _convinced_ they are, `prefetch_related` is about the most intensively Python part of the ORM convenience IME) operations, here's another example query, this time using a single `annotate` to add multiple extra values, and then a `values` call to pull those + some actual fields out.

baseline, main:
```
In [4]: %timeit User.objects.annotate(t1=Substr('username', 1, 5), t2=Value(True), t3=Concat(Value('begin|'), F('email'), Value("|end"), output_field=CharField())).values('username', 't1', 't2', 't3', 'is_active', 'is_staff', 'is_superuser')
188 µs ± 1.69 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each

In [5]: %timeit User.objects.annotate(t1=Substr('username', 1, 5), t2=Value(True), t3=Concat(Value('begin|'), F('email'), Value("|end"), output_field=CharField())).values('username', 't1', 't2', 't3', 'is_active', 'is_staff', 'is_superuser')
188 µs ± 1.46 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [6]: %timeit tuple(User.objects.annotate(t1=Substr('username', 1, 5), t2=Value(True), t3=Concat(Value('begin|'), F('email'), Value("|end"), output_field=CharField())).values('username', 't1', 't2', 't3', 'is_active', 'is_staff', 'is_superuser'))
4.31 ms ± 61.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [7]: %timeit tuple(User.objects.annotate(t1=Substr('username', 1, 5), t2=Value(True), t3=Concat(Value('begin|'), F('email'), Value("|end"), output_field=CharField())).values('username', 't1', 't2', 't3', 'is_active', 'is_staff', 'is_superuser'))
3.95 ms ± 74.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

After:
```
In [4]: %timeit User.objects.annotate(t1=Substr('username', 1, 5), t2=Value(True), t3=Concat(Value('begin|'), F('email'), Value("|end"), output_field=CharField())).values('username', 't1', 't2', 't3', 'is_active', 'is_staff', 'is_superuser')
183 µs ± 3.07 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [5]: %timeit tuple(User.objects.annotate(t1=Substr('username', 1, 5), t2=Value(True), t3=Concat(Value('begin|'), F('email'), Value("|end"), output_field=CharField())).values('username', 't1', 't2', 't3', 'is_active', 'is_staff', 'is_superuser'))
3.99 ms ± 71.7 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

Superficially, might be better, might be worse. Certainly that one too appears to be in the same ballpark. It would never actively benefit by much, because it's only **1** `Query` and **2** `clone()` calls.
It changes the following newly-class-attributes into instance ones during the course of execution: `default_cols`, `deferred_loading`, `filter_is_sticky`, `select`, `select_related`, `values_select`, `annotation_select_mask`, `_annotations_select_cache`, `extra_select_mask`, `_extra_select_cache`